FadeManager
It works but not in between scenes. My functions to do it between scenes
doesn't seem to work. Therefore calling it just before and after the
scene is loaded seems to be a better option.

//Fade out
FadeManager.Fade(true, 1.25f or another float number)

//Scene Load Script

//Fade in
FadeManager.Fade(false, 3f or another float number)