Update - Oni(NotSpawned) and LevelLoader
Made it so when player dies Oni that where not spawn by system will not
delete themself instead just go back to be 100% health if want change
that i can. Also, Create LevelLoader Script. It is Singleton ( If no wat
that mean good). It mean only one ever in the game so you can do find
and find it in the game when ever you need it. Right now script is
create but not add into the game yet. So dont try call it just yet. It
has one function.  LoadLevel- LoadLevel take in string and int. You can
give it a name of the level or index of the level( Put index of level
just be safe). If you have name and no index that find if got index but
not name that also fine. If have no idea and calling this function
sending it nothing .... HA i already plan for that and it will do
nothing. But Warning if you send in a name or index that does not work
it will error or crash.